Start of Clans overhaul...
This is long overdue, and now is the time.
- Switch to C99 size-based types and bool
- Don't pack in-memory structs
- Start of serialization/deserialization code

At this moment, while this builds, it won't work with any of the
data files, so it's useless.

Hopefully I can get it working again this week.
Goodbye _INT16, BOOL, __BOOL, WORD, DWORD, long, short... ahhh...

+++ src/doors/clans-src/src/crc.c
@@ -4,10 +4,10 @@
 
 #include "defines.h"
 
-long CRCValue(void *Data, int DataSize)
+int32_t CRCValue(void *Data, int DataSize)
 {
 	char *p;
-	long CRC = 0;
+	int32_t CRC = 0;
 
 	p = (char *)Data;
 
@@ -19,7 +19,7 @@ long CRCValue(void *Data, int DataSize)
 	return CRC;
 }
 
-int CheckCRC(void *Data, int Size, long CRC)
+int CheckCRC(void *Data, int Size, int32_t CRC)
 {
 	return (CRCValue(Data, Size) == CRC);
 }
